Renamed files for Sin_2chiGrouping plugin to be consistent with other filenames
