+++ src/pydidas_plugins/proc_plugins/sin_2chi_grouping.py
